+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes
 
 ### Unreleased
 
+* 2017-11-21 - Setting to remove some course nodes - Thanks to Alain Bolli.
 * 2017-11-08 - Updated travis.yml to use newer node version for fixing TravisCI error.
 
 ### v3.2-r15